requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dj-rest-auth[with_social]
drf-spectacular
gunicorn
whitenoise
django-fsm